测试所写功能并修改

+++ src/main/db/qz_setting_update_function_2019-7-24/qz_setting_related.sql
@@ -5,7 +5,7 @@ ALTER TABLE `db_keyword`.`t_qz_setting` ADD COLUMN `fOperatorStatus` TINYINT(1) 
 UPDATE t_qz_setting SET fRenewalStatus = 0, fStatus = 0 WHERE fStatus = 3;
 
 # 暴涨暴跌不仅需要涨幅指数大于小于指定值，还需要一周增长词数大于小于指定值
-ALTER TABLE `db_keyword`.`t_qz_keyword_rank_info` ADD COLUMN `fOneWeekDifference` TINYINT(1) DEFAULT 0 COMMENT '一周top10差值' AFTER `fTodayDifference`;
+ALTER TABLE `db_keyword`.`t_qz_keyword_rank_info` ADD COLUMN `fOneWeekDifference` INT(11) DEFAULT 0 COMMENT '一周top10差值' AFTER `fTodayDifference`;
 
 # 一周涨幅词数，配置在t_config 表
 INSERT INTO t_config VALUES('QZSettingKeywordRank', 'OneWeekDiff', '10');
